requirements.txt
graphene
graphene-sqlalchemy
SQLAlchemy
Flask
Flask-SQLAlchemy
Flask-GraphQL
python-dotenv
flask_jwt_extended
graphene_sqlalchemy_filter
argon2-cffi